Speed up transaction dialog improvements
