+++ WalletWasabi.Fluent/Views/Dialogs/SpeedUpTransactionDialogView.axaml.cs
@@ -3,9 +3,9 @@ using Avalonia.Markup.Xaml;
 
 namespace WalletWasabi.Fluent.Views.Dialogs;
 
-public class BoostTransactionDialogView : UserControl
+public class SpeedUpTransactionDialogView : UserControl
 {
-	public BoostTransactionDialogView()
+	public SpeedUpTransactionDialogView()
 	{
 		InitializeComponent();
 	}
